Add initial Baker site
This is just a prototype landing page, but we'll need to replace it with something more maintainable in the future. The legal status of using Baker is still...pretty odd. It's freeware, but am I legally allowed to redistribute my changes to freeware? (The sensible answer is "DUH! Look at the modding community!"...but when you deal with law, common senes flies out the window.)

The Baker system is actually pretty flawed, so I'm either going to rewrite it or switch the content out for some other framework with a "proper" license.